feat: fix tests, move more utils
