+++ packages/eslint-plugin/src/util/getWrappingFixer.ts
@@ -4,7 +4,6 @@ import {
   ASTUtils,
   TSESTree,
 } from '@typescript-eslint/experimental-utils';
-import { SourceCode } from '@typescript-eslint/experimental-utils/src/ts-eslint';
 
 interface WrappingFixerParams {
   /** Source code. */
@@ -135,7 +134,7 @@ function isWeakPrecedenceParent(node: TSESTree.Node): boolean {
  */
 function isMissingSemicolonBefore(
   node: TSESTree.Node,
-  sourceCode: SourceCode,
+  sourceCode: TSESLint.SourceCode,
 ): boolean {
   for (;;) {
     const parent = node.parent!;
