+++ packages/type-utils/README.md
@@ -1,6 +1,6 @@
 <h1 align="center">Type utils for ESLint Plugins</h1>
 
-<p align="center">Type utilities for working with TypeScript + ESLint together.</p>
+<p align="center">Type utilities for working with TypeScript within ESLint rules.</p>
 
 <p align="center">
     <img src="https://github.com/typescript-eslint/typescript-eslint/workflows/CI/badge.svg" alt="CI" />
@@ -8,19 +8,7 @@
     <a href="https://www.npmjs.com/package/@typescript-eslint/type-utils"><img src="https://img.shields.io/npm/dm/@typescript-eslint/type-utils.svg?style=flat-square" alt="NPM Downloads" /></a>
 </p>
 
-## Note
-
-This package has inherited its version number from the `@typescript-eslint` project.
-Meaning that even though this package is `2.x.y`, you shouldn't expect 100% stability between minor version bumps.
-i.e. treat it as a `0.x.y` package.
-
-Feel free to use it now, and let us know what utilities you need or send us PRs with utilities you build on top of it.
-
-## Exports
-
-| Name                                | Description                                           |
-| ----------------------------------- | ----------------------------------------------------- |
-| [`ESLintUtils`](./src/eslint-utils) | Tools for creating ESLint type rules with TypeScript. |
+This utilities in this package are separated from `@typescript-eslint/experimental-utils` so that that package does not require a dependency on `typescript`.
 
 ## Contributing
