--- packages/type-utils/src/eslint-utils/index.ts
@@ -1,2 +1,0 @@
-export * from './isTypeReadonly';
-export * from './propertyTypes';
