+++ packages/type-utils/src/index.ts
@@ -1,1 +1,14 @@
-export * as ESLintTypeUtils from './eslint-utils';
+export * from './containsAllTypesByName';
+export * from './getConstrainedTypeAtLocation';
+export * from './getContextualType';
+export * from './getDeclaration';
+export * from './getSourceFileOfNode';
+export * from './getTokenAtPosition';
+export * from './getTypeArguments';
+export * from './getTypeName';
+export * from './isTypeReadonly';
+export * from './isUnsafeAssignment';
+export * from './predicates';
+export * from './propertyTypes';
+export * from './requiresQuoting';
+export * from './typeFlagUtils';
