+++ packages/type-utils/src/isTypeReadonly.ts
