+++ packages/type-utils/src/predicates.ts
@@ -1,0 +1,152 @@
+import debug from 'debug';
+import { unionTypeParts } from 'tsutils';
+import * as ts from 'typescript';
+import { getTypeArguments } from './getTypeArguments';
+import { getTypeFlags, isTypeFlagSet } from './typeFlagUtils';
+
+const log = debug('typescript-eslint:eslint-plugin:utils:types');
+
+/**
+ * Checks if the given type is (or accepts) nullable
+ * @param isReceiver true if the type is a receiving type (i.e. the type of a called function's parameter)
+ */
+export function isNullableType(
+  type: ts.Type,
+  {
+    isReceiver = false,
+    allowUndefined = true,
+  }: { isReceiver?: boolean; allowUndefined?: boolean } = {},
+): boolean {
+  const flags = getTypeFlags(type);
+
+  if (isReceiver && flags & (ts.TypeFlags.Any | ts.TypeFlags.Unknown)) {
+    return true;
+  }
+
+  if (allowUndefined) {
+    return (flags & (ts.TypeFlags.Null | ts.TypeFlags.Undefined)) !== 0;
+  } else {
+    return (flags & ts.TypeFlags.Null) !== 0;
+  }
+}
+
+/**
+ * Checks if the given type is either an array type,
+ * or a union made up solely of array types.
+ */
+export function isTypeArrayTypeOrUnionOfArrayTypes(
+  type: ts.Type,
+  checker: ts.TypeChecker,
+): boolean {
+  for (const t of unionTypeParts(type)) {
+    if (!checker.isArrayType(t)) {
+      return false;
+    }
+  }
+
+  return true;
+}
+
+/**
+ * @returns true if the type is `unknown`
+ */
+export function isTypeUnknownType(type: ts.Type): boolean {
+  return isTypeFlagSet(type, ts.TypeFlags.Unknown);
+}
+
+/**
+ * @returns true if the type is `any`
+ */
+export function isTypeAnyType(type: ts.Type): boolean {
+  if (isTypeFlagSet(type, ts.TypeFlags.Any)) {
+    if (type.intrinsicName === 'error') {
+      log('Found an "error" any type');
+    }
+    return true;
+  }
+  return false;
+}
+
+/**
+ * @returns true if the type is `any[]`
+ */
+export function isTypeAnyArrayType(
+  type: ts.Type,
+  checker: ts.TypeChecker,
+): boolean {
+  return (
+    checker.isArrayType(type) &&
+    isTypeAnyType(
+      // getTypeArguments was only added in TS3.7
+      getTypeArguments(type, checker)[0],
+    )
+  );
+}
+
+/**
+ * @returns true if the type is `unknown[]`
+ */
+export function isTypeUnknownArrayType(
+  type: ts.Type,
+  checker: ts.TypeChecker,
+): boolean {
+  return (
+    checker.isArrayType(type) &&
+    isTypeUnknownType(
+      // getTypeArguments was only added in TS3.7
+      getTypeArguments(type, checker)[0],
+    )
+  );
+}
+
+export const enum AnyType {
+  Any,
+  AnyArray,
+  Safe,
+}
+/**
+ * @returns `AnyType.Any` if the type is `any`, `AnyType.AnyArray` if the type is `any[]` or `readonly any[]`,
+ *          otherwise it returns `AnyType.Safe`.
+ */
+export function isAnyOrAnyArrayTypeDiscriminated(
+  node: ts.Node,
+  checker: ts.TypeChecker,
+): AnyType {
+  const type = checker.getTypeAtLocation(node);
+  if (isTypeAnyType(type)) {
+    return AnyType.Any;
+  }
+  if (isTypeAnyArrayType(type, checker)) {
+    return AnyType.AnyArray;
+  }
+  return AnyType.Safe;
+}
+
+/**
+ * @returns Whether a type is an instance of the parent type, including for the parent's base types.
+ */
+export function typeIsOrHasBaseType(
+  type: ts.Type,
+  parentType: ts.Type,
+): boolean {
+  const parentSymbol = parentType.getSymbol();
+  if (!type.getSymbol() || !parentSymbol) {
+    return false;
+  }
+
+  const typeAndBaseTypes = [type];
+  const ancestorTypes = type.getBaseTypes();
+
+  if (ancestorTypes) {
+    typeAndBaseTypes.push(...ancestorTypes);
+  }
+
+  for (const baseType of typeAndBaseTypes) {
+    const baseSymbol = baseType.getSymbol();
+    if (baseSymbol && baseSymbol.name === parentSymbol.name) {
+      return true;
+    }
+  }
+
+  return false;
+}
